backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
daphne==4.1.2
Django==4.2.13
django-cors-headers==4.3.1
django-extensions==3.2.3
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-two-factor-auth==1.16.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
msgpack==1.0.8
msgpack-python==0.5.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.40
pillow==10.3.0
platformdirs==4.2.2
psycopg==3.1.19
psycopg-binary==3.1.19
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypng==0.20220715.0
qrcode==7.4.2
redis==5.0.6
requests==2.32.3
service-identity==24.1.0
setuptools==70.2.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
whitenoise==6.7.0
zope.interface==6.4.post2
